Add SML migration feature

+++ generic-server/src/main/java/org/holodeckb2b/bdxr/smp/server/ui/controllers/ParticipantsViewController.java
@@ -29,6 +29,7 @@ import org.holodeckb2b.bdxr.smp.server.db.repos.ParticipantRepository;
 import org.holodeckb2b.bdxr.smp.server.db.repos.ServiceMetadataBindingRepository;
 import org.holodeckb2b.bdxr.smp.server.db.repos.ServiceMetadataTemplateRepository;
 import org.holodeckb2b.bdxr.smp.server.svc.DirectoryIntegrationService;
+import org.holodeckb2b.bdxr.smp.server.svc.SMLException;
 import org.holodeckb2b.bdxr.smp.server.svc.SMLIntegrationService;
 import org.holodeckb2b.commons.util.Utils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -97,7 +98,8 @@ public class ParticipantsViewController {
 		ParticipantE p = null;
 		try {
 			p = participants.getReferenceById(oid);
-			if (smlService.isSMLIntegrationAvailable() && p.registeredInSML()) {
+			if (smlService.isSMLIntegrationAvailable() && p.registeredInSML() 
+				&& Utils.isNullOrEmpty(p.getMigrationCode())) {
 				if (p.publishedInDirectory())
 					directoryService.removeParticipantInfo(p);
 				smlService.unregisterParticipant(p);
@@ -117,75 +119,86 @@ public class ParticipantsViewController {
 	}
 
 	@PostMapping(value = "/update", params = { "save" })
-	public String saveParticipant(@ModelAttribute(P_ATTR) @Valid ParticipantE input, BindingResult br, Model m, HttpSession s) {
-		
-		ParticipantE prev = (ParticipantE) s.getAttribute(P_ATTR);
-		ParticipantE current = participants.findByIdentifier(input.getId()); 
-		if (!br.hasErrors() && current != null && current.getOid() != prev.getOid()) {
+	public String saveParticipant(@ModelAttribute(P_ATTR) @Valid ParticipantE input, BindingResult br, Model m, HttpSession s) {		
+		ParticipantE stored = participants.findByIdentifier(input.getId()); 
+		if (!br.hasErrors() && stored != null && stored.getOid() != input.getOid()) {
 			br.rejectValue("id.value", "ID_EXISTS", "There already exists another Participant with the same Identifier");
 			br.rejectValue("id.scheme", "ID_EXISTS");
 		}
 		if (input.publishedInDirectory()) {
-			if (!input.registeredInSML()) {
+			if (input.getIsRegisteredSML() != null && !input.getIsRegisteredSML()) 
 				br.rejectValue("publishedInDirectory", "SML_REG_REQD", 
 								"To publish the business information in the directory, the participant must also be registered in the SML");
-				input.setIsRegisteredSML(prev.registeredInSML());
-			}
 			if (Utils.isNullOrEmpty(input.getName()))
 				br.rejectValue("name", "NAME_REQD", "Name is required when publishing to the directory");
 			if (Utils.isNullOrEmpty(input.getCountry()))
-					br.rejectValue("country", "COUNTRY_REQD", "Country is required when publishing to the directory");
+				br.rejectValue("country", "COUNTRY_REQD", "Country is required when publishing to the directory");
 		}
 
 		if (br.hasErrors()) 
 			return toForm(m, s);
 
-		boolean smlUpdRequired = smlService.isSMLIntegrationAvailable() &&
-								( (current != null && current.registeredInSML() != input.registeredInSML())
-								|| (current == null && input.registeredInSML()) );
-		ParticipantE updated = updateSessionBasics(input, s); 
+		boolean wasSMLRegistered = stored != null && stored.registeredInSML();
+		boolean wasPublished = stored != null && stored.publishedInDirectory();
+		String storedMigrationCode = stored != null ? stored.getMigrationCode() : null;
 		try {
-			// Save data
-			updated = participants.save(updated);
+			// First update the Participant data in database,			
+			stored = participants.save(getUpdatedParticipant(input, s));
 			
-			// Update SML registration if needed
-			if (smlUpdRequired)
+			// If SML integration is enabled, check if an update in the SML is needed
+			if (smlService.isSMLIntegrationAvailable()) {
 				try {
-					if (updated.registeredInSML())
-						smlService.registerParticipant(updated);
-					else
-						smlService.unregisterParticipant(updated);
+					if (wasSMLRegistered && !stored.registeredInSML()) 
+						smlService.unregisterParticipant(stored);
+					else if (!wasSMLRegistered && stored.registeredInSML()) {
+						if (Utils.isNullOrEmpty(stored.getMigrationCode()))
+							smlService.registerParticipant(stored);
+						else {
+							smlService.migrateParticipant(stored);
+							stored.setMigrationCode(null);
+							stored = participants.save(stored);
+						}								
+					} else if (wasSMLRegistered && Utils.isNullOrEmpty(storedMigrationCode) 
+							&& !Utils.isNullOrEmpty(stored.getMigrationCode())) {
+						// A migration code has been set and therefore needs to be registered in the SML
+						smlService.registerMigrationCode(stored);
+					}
 				} catch (Exception smlUpdateFailed) {
-					updated.setIsRegisteredSML(!updated.registeredInSML());
-					updated = participants.save(updated);
+					stored.setIsRegisteredSML(wasSMLRegistered);
+					stored.setMigrationCode(storedMigrationCode);
+					stored = participants.save(stored);
 					m.addAttribute("errorMessage", "There was an error updating the Participant's registration in the SML ("
-									+ smlUpdateFailed.getMessage() + ")");
+									+ Utils.getRootCause(smlUpdateFailed).getMessage() + ")");
 				}		
-			if (updated.publishedInDirectory())
+			}
+			// If directory integration is enabled, check if an update in the directory is needed
+			if (directoryService.isDirectoryIntegrationAvailable()) {
 				try {
-					directoryService.publishParticipantInfo(updated);
+					if (stored.publishedInDirectory())
+						directoryService.publishParticipantInfo(stored);
+					else if (wasPublished)
+						directoryService.removeParticipantInfo(stored);					
 				} catch (Exception dirUpdatedFailed) {
-					m.addAttribute("errorMessage", "There was an error publishing the Participant's info to the directory ("
-									+ dirUpdatedFailed.getMessage() + ")");
-				}
-			else if (current != null && current.publishedInDirectory()) 
-				try {
-					directoryService.removeParticipantInfo(updated);
-				} catch (Exception dirUpdatedFailed) {
-					updated.setPublishedInDirectory(true);
-					updated = participants.save(updated);				
-					m.addAttribute("errorMessage", "There was an error removing the Participant's info from the directory ("
-							+ dirUpdatedFailed.getMessage() + ")");
+					m.addAttribute("errorMessage", "There was an error " 
+									+ (stored.publishedInDirectory() ? "publishing" : "removing") 
+									+ " the Participant's info to the directory ("
+									+ Utils.getRootCause(dirUpdatedFailed).getMessage() + ")");
+
+					stored.setPublishedInDirectory(!stored.publishedInDirectory());
+					stored = participants.save(stored);				
+					
 					return toForm(m, s);					
 				}			
+			}
 		} catch (Exception updateFailure) {
-			log.error("Failed to update the Participant (PID={}) meta-data : {}", prev.getOid(), Utils.getExceptionTrace(updateFailure));
+			log.error("Failed to update the Participant (PID={}) meta-data : {}", input.getId().toString(),
+						Utils.getExceptionTrace(updateFailure));
 			m.addAttribute("errorMessage", "There was an error updating the Participant's info");
 		}
 		
 		if (m.getAttribute("errorMessage") != null) {
-			m.addAttribute(P_ATTR, updated);
-			s.setAttribute(P_ATTR, updated);
+			m.addAttribute(P_ATTR, stored);
+			s.setAttribute(P_ATTR, stored);
 			return toForm(m, s);
 		} else		
 			return "redirect:/participants";
@@ -193,14 +206,34 @@ public class ParticipantsViewController {
 
 	@PostMapping(value = "/update", params = {"addBinding", "template2add"})
 	public String addService(@ModelAttribute(P_ATTR) ParticipantE input, Model m, HttpSession s, @RequestParam("template2add") Long svc2add) {
-		ParticipantE p = updateSessionBasics(input, s);
+		ParticipantE p = (ParticipantE) s.getAttribute(P_ATTR);
 		p.getBindings().add(new ServiceMetadataBindingE(p, templates.getById(svc2add)));
 		return toForm(m, s);
 	}
 
 	@PostMapping(value = "/update", params = {"removeBinding"})
 	public String removeService(@ModelAttribute(P_ATTR) ParticipantE input, Model m, HttpSession s, @RequestParam("removeBinding") Long row) {
-		updateSessionBasics(input, s).getBindings().remove(row.intValue());
+		((ParticipantE) s.getAttribute(P_ATTR)).getBindings().remove(row.intValue());
+		return toForm(m, s);
+	}
+	
+	@PostMapping(value = "/update", params = {"cancelMigration"})
+	public String cancelMigration(@ModelAttribute(P_ATTR) ParticipantE input, Model m, HttpSession s) {
+		ParticipantE stored = participants.findByIdentifier(input.getId());
+		try {
+			smlService.migrateParticipant(stored);
+			stored.setMigrationCode(null);
+			stored = participants.save(stored);	
+			input.setMigrationCode(null);
+			((ParticipantE) s.getAttribute(P_ATTR)).setMigrationCode(null);
+		} catch (SMLException migrateFailed) {
+			m.addAttribute("errorMessage", 
+					"Could not cancel migration, probably because registration has already moved to other SMP. Error: " 
+					+ Utils.getRootCause(migrateFailed).getMessage());
+		}
+		// Because the checkbox for SML registration is disabled, its value is not set in the model and we need to set
+		// it again
+		input.setIsRegisteredSML(true);
 		return toForm(m, s);
 	}
 
@@ -212,24 +245,38 @@ public class ParticipantsViewController {
 		else
 			m.addAttribute(P_ATTR, sp);
 
+		m.addAttribute("isMigrating", !Utils.isNullOrEmpty(sp.getMigrationCode()));
 		m.addAttribute("availableSMT",  templates.findAll().stream().filter(smt -> sp.getBindings().parallelStream()
 												.noneMatch(b -> b.getTemplate().equals(smt)))
 												.toList());
  		return "admin-ui/participant_form";
 	}
 
-	private ParticipantE updateSessionBasics(ParticipantE input, HttpSession s) {
-		ParticipantE stored = (ParticipantE) s.getAttribute(P_ATTR);
+	/**
+	 * Gets the updated data on the Particiant by combining the basic fields from the Model and the bound SMT from the
+	 * session object.
+	 * 
+	 * @param input	the meta-data from the model
+	 * @param s		the meta-data from the session
+	 * @return	the combined data
+	 */
+	private ParticipantE getUpdatedParticipant(ParticipantE input, HttpSession s) {
+		ParticipantE updated = (ParticipantE) s.getAttribute(P_ATTR);
 		if (input != null) {
-			stored.setId(input.getId());
-			stored.setName(input.getName());
-			stored.setCountry(input.getCountry());
-			stored.setAddressInfo(input.getAddressInfo());
-			stored.setContactInfo(input.getContactInfo());
-			stored.setIsRegisteredSML(input.getIsRegisteredSML());
-			stored.setPublishedInDirectory(input.getPublishedInDirectory());
+			updated.setId(input.getId());
+			updated.setName(input.getName());
+			updated.setCountry(input.getCountry());
+			updated.setAddressInfo(input.getAddressInfo());
+			updated.setContactInfo(input.getContactInfo());
+			updated.setPublishedInDirectory(input.getPublishedInDirectory());
+			updated.setFirstRegistration(input.getFirstRegistration());
+			updated.setAdditionalIds(input.getAdditionalIds());
+			if (Utils.isNullOrEmpty(updated.getMigrationCode())) {
+				updated.setIsRegisteredSML(input.getIsRegisteredSML());
+				updated.setMigrationCode(input.getMigrationCode());
+			}
 		}
-		return stored;
+		return updated;
 	}
 }
